requirements-ml.txt
prophet==1.3.0
scikit-learn==1.8.0
matplotlib==3.10.9
pandas==2.3.3
sqlalchemy==2.0.49
psycopg2-binary==2.9.12
python-dotenv==1.0.1